pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoice_bot"
version = "0.1.0"
description = "Headless browser automation for invoice uploads"
requires-python = ">=3.11"
dependencies = [
    "playwright==1.52.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "rich>=13.5.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "ruff>=0.0.272",
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "UP", "SIM", "ARG"]
ignore = []